+++ src/components/other/SearchBox.tsx
@@ -2,16 +2,19 @@ import { HiOutlineSearch } from "react-icons/hi";
 
 export type SearchBoxProps = {
   className?: string;
+  iconClassName?: string;
+  inputClassName?: string;
 };
 
 const SearchBox = (props: SearchBoxProps) => {
-  const className = props.className ?? "";
+  const { className, iconClassName, inputClassName } = props;
+
   return (
     <>
       <div className={`search-box ${className}`}>
-        <HiOutlineSearch className="search-icon" />
+        <HiOutlineSearch className={`search-icon ${iconClassName}`} />
 
-        <input placeholder="Search..." />
+        <input placeholder="Search..." className={`${inputClassName}`} />
       </div>
     </>
   );